batch: Extract baseline edit fallback
Batch merge owns structural placement, ambiguity previews, and exact
baseline-coordinate fallback logic.

Baseline-coordinate fallback has a separate responsibility: it proves
recorded baseline references and streams non-overlapping edits when LCS
anchors cannot place replacement units.

This commit moves that fallback into batch.baseline_edits and rewires
batch.merge to call the new owner for replacement-origin choices,
origin-missing checks, and exact baseline edits.

Tests will follow the new module boundary and remove the remaining
white-box import from batch.merge.
